Reorganise code for kernel websocket handler
The inheritance stack was making this harder to follow.
